requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.9
click==8.1.3
idna==3.3
joblib==1.1.0
LeIA==0.1.2
nltk==3.7
numpy==1.22.4
python-dotenv==0.20.0
regex==2022.4.24
requests==2.27.1
tqdm==4.64.0
urllib3==1.26.9
